% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_fia_plots.R
\name{calculate_relative_matrix}
\alias{calculate_relative_matrix}
\title{Title}
\usage{
calculate_relative_matrix(mat, point.threshold = 0)
}
\arguments{
\item{point.threshold}{}
}
\description{
Title
}
