% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mapping_project.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Calculate the weights of each grid cell in a CFI plot}
\usage{
calculate_weights(idw.power = 3, n.subplots = 4)
}
\arguments{
\item{idw.power}{The weight to use for IDW}

\item{n.subplots}{The number of subplots in each CFI plot (4 for FIA)}
}
\value{
A matrix of weights
}
\description{
If idw.power is one of 0.25, 0.5, 1, 2, or 3 and n.subplots = 4 then
precomputed weights are loaded.
}
