% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_tools.R
\name{check_numeric}
\alias{check_numeric}
\title{Check local numeric data against remote numeric data}
\usage{
check_numeric(local, remote, column)
}
\arguments{
\item{column}{}
}
\value{
A sina or violin plot comparing the local and remote data
}
\description{
Check local numeric data against remote numeric data
}
