% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_practices_functions.R
\name{checkpoint_directory}
\alias{checkpoint_directory}
\title{Maintain a list of previously processed files within a directory}
\usage{
checkpoint_directory(path, pattern = NULL, qaqc.file = NULL)
}
\arguments{
\item{path}{A character vector of a directory containing files}

\item{pattern}{A pattern to match within the file names}

\item{qaqc.file}{The destination file of the data after QAQC/processing}
}
\description{
Maintain a list of previously processed files within a directory
}
