% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mapping_project.R
\name{collapse_matrices}
\alias{collapse_matrices}
\title{Collapse a list-column of matrices into a single matrix}
\usage{
collapse_matrices(matrix.col, fun = cbind)
}
\arguments{
\item{matrix.col}{A list column of matrices}

\item{fun}{One of \code{rbind} or \code{cbind} used to combine the matrices}
}
\value{
A list of matrices
}
\description{
Collapse a list-column of matrices into a single matrix
}
