% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{combine_se}
\alias{combine_se}
\title{Combine Standard Errors}
\usage{
combine_se(mu, se, n)
}
\arguments{
\item{mu}{A vector of two means}

\item{se}{A vector of two standard errors}

\item{n}{A vector of two group sizes}
}
\description{
Combine Standard Errors
}
