% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimadzu_functions.R
\name{correct_shimadzu}
\alias{correct_shimadzu}
\title{Fix Shimadzu transcription errors}
\usage{
correct_shimadzu(.data, sourceFile.string, site, sampleLocation,
  sampleDate, sampleTime, error.column, replacement.value)
}
\arguments{
\item{replacement.value}{}
}
\description{
Fix Shimadzu transcription errors
}
