% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mapping_project.R
\name{get_dominant_species}
\alias{get_dominant_species}
\title{Return a single dominant species for each plot}
\usage{
get_dominant_species(mat, point.threshold = 0.2, plot.threshold = 0.2)
}
\arguments{
\item{mat}{A matrix of relative values}

\item{point.threshold}{A threshold specifying what proportion of the relative
measure at given point a species has represent to be considered present.
Passed on to calculate_relative_matrix}

\item{plot.threshold}{A threshold specifying what proportion of a plot a
species must be present in to be considered dominant}
}
\value{
A character vector of the single species that makes up the largest
  share of the measure in a plot, or 'mixed' if no species makes up more than
  \code{plot.threshold}
}
\description{
Return a single dominant species for each plot
}
\seealso{
Other plot classifiers: \code{\link{classify_plot}},
  \code{\link{get_species_presence}},
  \code{\link{get_species_proportion}}
}
\concept{plot classifiers}
