% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_practices_functions.R
\name{get_qaqc_file}
\alias{get_qaqc_file}
\title{Extract the qaqc filename from __procesed_files.txt}
\usage{
get_qaqc_file(x)
}
\arguments{
\item{x}{The path to __processed_files.txt}
}
\description{
Extract the qaqc filename from __procesed_files.txt
}
