% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_fia_plots.R
\name{get_species_presence}
\alias{get_species_presence}
\title{Title}
\usage{
get_species_presence(mat, point.threshold = 0, plot.threshold = 0.2)
}
\arguments{
\item{plot.threshold}{}
}
\description{
Title
}
