% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mapping_project.R
\name{get_species_proportion}
\alias{get_species_proportion}
\title{Get the mean relative measure of species in a plot}
\usage{
get_species_proportion(mat, point.threshold = 0.2,
  plot.threshold = 0.2)
}
\arguments{
\item{mat}{A matrix of relative values}

\item{point.threshold}{A threshold specifying what proportion of the relative
measure at given point a species has represent to be considered 'present'.
Passed on to calculate_relative_matrix}

\item{plot.threshold}{A threshold specifying what proportion of a plot a
species must be present in to be considered present}
}
\value{
A tibble of n-columns with the plot-level relative measure of the
  species whose mean relative measure across the plot is greater than
  plot.threshold
}
\description{
Get the mean relative measure of species in a plot
}
\seealso{
Other plot classifiers: \code{\link{classify_plot}},
  \code{\link{get_dominant_species}},
  \code{\link{get_species_presence}}
}
\concept{plot classifiers}
