% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory_functions.R
\name{importance_value}
\alias{importance_value}
\title{Calculate Modified importance value}
\usage{
importance_value(basal.area, prevalence)
}
\arguments{
\item{basal.area}{a numeric vector of basal areas}

\item{prevalence}{a numeric vector of prevalence}
}
\value{
A numeric vector
}
\description{
Calculates a modified importance value using relative dominance (basal area)
and relative density (relative prevalence) rescaled to fall between 0 and 1.
}
\details{
Importance values were originally described in Curtis and MacIntosh (1951).
An Upland Forest Continuum in the Prarie-Forest Border Region of Wisconsin.
}
\examples{

}
