% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_tools.R
\name{insert_data}
\alias{insert_data}
\title{Title}
\usage{
insert_data(local, remote, verbose = TRUE, ignore.structure = FALSE)
}
\arguments{
\item{ignore.structure}{}
}
\description{
Title
}
