% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mapping_project.R
\name{interpolate_plot}
\alias{interpolate_plot}
\title{Interpolate a given measure across the entire CFI plot}
\usage{
interpolate_plot(measure, subgroup, ...)
}
\arguments{
\item{measure}{A numeric vector}

\item{subgroup}{A grouping variable supplied as length 1 or same length as
measure}

\item{...}{Other parameters passed on to \code{\link{calculate_weights}}}
}
\value{
A list of matrices equal in length to measure
}
\description{
Interpolate a given measure across the entire CFI plot
}
