% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{interval_join}
\alias{interval_join}
\title{Title}
\usage{
interval_join(x, y, instant, interval, by = NULL, join,
  step = "15 minutes")
}
\arguments{
\item{step}{}
}
\description{
Title
}
