% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{parse_excel_date}
\alias{parse_excel_date}
\title{Parse a date time that has been opened in Excel}
\usage{
parse_excel_date(x)
}
\arguments{
\item{x}{}
}
\description{
Parse a date time that has been opened in Excel
}
