% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{pretty_png}
\alias{pretty_png}
\title{Export a png from base graphics}
\usage{
pretty_png(file.name, plot.code, res = 300, width = 3000,
  height = 2000, ...)
}
\arguments{
\item{file.name}{The output location}

\item{plot.code}{A function or code chunk to create the figure}

\item{res}{Figure resolution}

\item{width}{Figure width in pixels}

\item{height}{Figure height in pixels}

\item{...}{Additional par() settings}
}
\description{
By default Liberation Narrow is used as the font family, and the axis text
title and labels are scaled to match the output resolution. Axis text is
rotated to always be horizontal (las = 1)
}
