% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_32bit_access.R
\name{r32_fetch_table}
\alias{r32_fetch_table}
\title{r32_fetch_table}
\usage{
r32_fetch_table(db.path, table.name)
}
\arguments{
\item{table.name}{}
}
\description{
Fetch a table from an Access database from 64-bit R.
}
