% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_32bit_access.R
\name{r32_list_tables}
\alias{r32_list_tables}
\title{r32_list_tables}
\usage{
r32_list_tables(db.path)
}
\arguments{
\item{db.path}{}
}
\description{
List tables in a 32-bit Access database from 64-bit R
}
