% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_32bit_access.R
\name{r32_source}
\alias{r32_source}
\title{Source a file using a 32-bit R process}
\usage{
r32_source(source.file)
}
\arguments{
\item{source.file}{}
}
\description{
Source a file using a 32-bit R process
}
