% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{range_test}
\alias{range_test}
\title{Title}
\usage{
range_test(data, observations, lower = NULL, upper = NULL,
  backfill = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
