% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{save_ggplot}
\alias{save_ggplot}
\title{Auto scale text in ggsave}
\usage{
save_ggplot(filename, plot = last_plot(), ...)
}
\arguments{
\item{...}{}
}
\description{
This function autoscales the fonts from X11() windows display at 96ppi to the
output resolution specified with dpi.
}
