% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{set_rprofile}
\alias{set_rprofile}
\title{Copy the supplied .Rprofile to your home directory}
\usage{
set_rprofile(overwrite = TRUE, view.only = FALSE)
}
\arguments{
\item{overwrite}{Logical indicating if an exisits .Rprofile should be
overwritten}

\item{view.only}{Logical indicating if .Rprofile should be printed in the
console instead of copied}
}
\description{
Copy the supplied .Rprofile to your home directory
}
