% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{set_treatment_period}
\alias{set_treatment_period}
\title{Set black ash study site treatment period}
\usage{
set_treatment_period(x, study = "pws")
}
\arguments{
\item{x}{Time series of sample data/times}

\item{study}{One of "pws", "eco", "planting"}
}
\value{
A factor with levels "Pre-treatment" and "Post-treatment"
}
\description{
Set black ash study site treatment period
}
