% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{spike_test}
\alias{spike_test}
\title{Title}
\usage{
spike_test(data, observations, n = 4, window = NULL, max.passes = 3,
  backfill = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
