% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mapping_project.R
\name{split_cfi_id}
\alias{split_cfi_id}
\alias{extract_cfi_id}
\title{Split a CFI plot id into constituent parts.}
\usage{
split_cfi_id(x)

extract_cfi_id(data, x)
}
\arguments{
\item{x}{A character vector, or for \code{extract_cfi_id} the name of a column}

\item{data}{The data containing the CFI column for \code{extract_cfi_id}}
}
\value{
\code{split_cfi_id} returns a character vector and
  \code{extract_cfi_id} runs \code{extract} on a column in a tibble
}
\description{
Split a CFI plot id into constituent parts.
}
