% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_practices_functions.R
\name{unprocess_directory}
\alias{unprocess_directory}
\title{Remove a directory of processed data}
\usage{
unprocess_directory(path)
}
\arguments{
\item{path}{The directory to 'unprocess'}
}
\description{
This will remove the __processed_files.txt file and delete the data from the
QAQC file specified in __processed_files.txt
}
