% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimadzu_functions.R
\name{validate_shimadzu}
\alias{validate_shimadzu}
\title{Validate sample sites and locations}
\usage{
validate_shimadzu(.data)
}
\arguments{
\item{.data}{}
}
\description{
Validate sample sites and locations
}
