% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_practices_functions.R
\name{write_qaqc}
\alias{write_qaqc}
\title{Append data to an existing QAQC file and create a back-up}
\usage{
write_qaqc(data, source.directory, ignore.names = FALSE)
}
\arguments{
\item{data}{The data to write}

\item{source.directory}{The directory of the raw files}

\item{ignore.names}{Logical indicating whether or not to check if the names
of the existing data and new data should match}

\item{check.duplicates}{Logical indicating whether the appended QAQC data
should check for duplicates of site and sample time}
}
\description{
Append data to an existing QAQC file and create a back-up
}
