% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mapping_project.R
\name{calculate_relative_matrix}
\alias{calculate_relative_matrix}
\title{Calculate the rowwise relative values of a matrix}
\usage{
calculate_relative_matrix(mat, point.threshold = 0.2)
}
\arguments{
\item{mat}{The matrix of raw values}

\item{point.threshold}{A threshold below which the value should be set to zero}
}
\value{
A matrix of the same dimensions as mat
}
\description{
If \code{point.threshold} is use values are not iteratively removed, just one
pass is performed and the relative values are not adjusted
}
