% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_fia_plots.R
\name{classify_plot}
\alias{classify_plot}
\title{Title}
\usage{
classify_plot(data, ..., classifier = get_dominant_species,
  classifier.options = NULL, idw_power = 3, use.parallel = FALSE)
}
\arguments{
\item{use.parallel}{}
}
\description{
Title
}
