% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_fia_plots.R
\name{get_dominant_species}
\alias{get_dominant_species}
\title{Title}
\usage{
get_dominant_species(mat, point.threshold = 0, plot.threshold = 0)
}
\arguments{
\item{plot.threshold}{}
}
\description{
Title
}
