% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_practices_functions.R
\name{write_qaqc}
\alias{write_qaqc}
\title{Append data to an existing QAQC file and create a back-up}
\usage{
write_qaqc(data, input.files, ignore.names = FALSE)
}
\arguments{
\item{data}{The data to write}

\item{input.files}{The files used in processing}

\item{ignore.names}{Logical indicating whether or not to check if the names
of the existing data and new data should match}
}
\description{
Append data to an existing QAQC file and create a back-up
}
